pyproject.toml
[project]
name = "cortex-mistral"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "mistralai>=1.8.2",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "pymupdf>=1.26.1",
    "pytesseract>=0.3.13",
    "requests>=2.32.4",
    "typing>=3.10.0.0",
]
